cleaned up the Storybook stories
