+++ storybook/stories/bar-chart/index.js
@@ -1,25 +1,15 @@
 import React from 'react'
-import { BarChart, Grid } from 'react-native-svg-charts'
+import { storiesOf } from '@storybook/react-native'
 
-class BarChartExample extends React.PureComponent {
+import Standard from './standard'
+import Horizontal from './horizontal'
+import Grouped from './grouped'
+import GroupedHorizontal from './horizontal-grouped'
+import ShowcaseCard from '../showcase-card'
 
-    render() {
-
-        const fill = 'rgb(134, 65, 244)'
-        const data   = [ 50, 10, 40, 95, -4, -24, null, 85, undefined, 0, 35, 53, -53, 24, 50, -20, -80 ]
-
-        return (
-            <BarChart
-                style={{ height: 200 }}
-                data={ data }
-                svg={{ fill }}
-                contentInset={{ top: 30, bottom: 30 }}
-            >
-                <Grid/>
-            </BarChart>
-        )
-    }
-
-}
-
-export default BarChartExample
+storiesOf('BarChart', module)
+    .addDecorator(getStory => <ShowcaseCard>{ getStory() }</ShowcaseCard>)
+    .add('Standard', () => <Standard/>)
+    .add('Grouped', () => <Grouped/>)
+    .add('Horizontal', () => <Horizontal/>)
+    .add('Horizontal - grouped', () => <GroupedHorizontal/>)
