+++ storybook/stories/line-chart/standard.js
@@ -1,0 +1,24 @@
+import React from 'react'
+import { LineChart, Grid } from 'react-native-svg-charts'
+
+class LineChartExample extends React.PureComponent {
+
+    render() {
+
+        const data = [ 50, 10, 40, 95, -4, -24, 85, 91, 35, 53, -53, 24, 50, -20, -80 ]
+
+        return (
+            <LineChart
+                style={{ height: 200 }}
+                data={ data }
+                svg={{ stroke: 'rgb(134, 65, 244)' }}
+                contentInset={{ top: 20, bottom: 20 }}
+            >
+                <Grid/>
+            </LineChart>
+        )
+    }
+
+}
+
+export default LineChartExample
