+++ storybook/stories/others/index.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react-native'
 
-import Regular from './standard'
-import Partial from './partial'
+import YMinMax from './y-min-max'
+import LayeredChart from './layered-charts'
 import ShowcaseCard from '../showcase-card'
 
-storiesOf('AreaChart', module)
+storiesOf('Others', module)
     .addDecorator(getStory => <ShowcaseCard>{ getStory() }</ShowcaseCard>)
-    .add('Standard', () => <Regular/>)
-    .add('Partial', () => <Partial/>)
+    .add('YMinMax', () => <YMinMax/>)
+    .add('Layered chart', () => <LayeredChart/>)
