+++ storybook/stories/pie-chart/index.js
@@ -1,33 +1,13 @@
 import React from 'react'
-import { PieChart } from 'react-native-svg-charts'
+import { storiesOf } from '@storybook/react-native'
 
-class PieChartExample extends React.PureComponent {
+import Standard from './standard'
+import WithCenteredText from './with-center-text'
+import WithLabels from './with-labels'
+import ShowcaseCard from '../showcase-card'
 
-    render() {
-
-        const data = [ 50, 10, 40, 95, -4, -24, 85, 91, 35, 53, -53, 24, 50, -20, -80 ]
-
-        const randomColor = () => ('#' + (Math.random() * 0xFFFFFF << 0).toString(16) + '000000').slice(0, 7)
-
-        const pieData = data
-            .filter(value => value > 0)
-            .map((value, index) => ({
-                value,
-                svg: {
-                    fill: randomColor(),
-                    onPress: () => console.log('press', index),
-                },
-                key: `pie-${index}`,
-            }))
-
-        return (
-            <PieChart
-                style={{ height: 200 }}
-                data={ pieData }
-            />
-        )
-    }
-
-}
-
-export default PieChartExample
+storiesOf('PieChart', module)
+    .addDecorator(getStory => <ShowcaseCard>{ getStory() }</ShowcaseCard>)
+    .add('Standard', () => <Standard/>)
+    .add('With centered text', () => <WithCenteredText/>)
+    .add('With labels', () => <WithLabels/>)
