+++ storybook/stories/progress-circle/gauge.js
